scripts/lexicon_requirements.txt
pyyaml
wordfreq
